Fix Linux CI QTest column type mismatch in timeout config test
invalidLimitsOrTimeoutIsRejectedAsInvalidConfigurationWithoutThrowing_data()
declares its "timeoutMs" column as qint64 (Qt's always-"long long"
alias), but the "timeout-above-sane-cap" row pushed
kMaxAllowedTimeout.count() + 1 without a cast.
std::chrono::milliseconds::rep is platform-defined: it is "long long"
under macOS/libc++ (matching qint64 exactly, so this passed locally),
but "long" under 64-bit Linux/glibc -- a distinct type from qint64
despite being the same width. QTest::addColumn/newRow require an exact
type match per column, so on Linux CI this silently registered a
mismatched metatype and QFETCH(qint64, timeoutMs) aborted with a fatal
"ASSERT: false" in qtestdata.cpp, failing the whole Unit tests job.

Fix: explicitly static_cast<qint64>(...) the pushed value so the
column's registered type is identical on every platform.

+++ tests/AssetNetworkFetcherTests.cpp
@@ -1453,9 +1453,18 @@ void AssetNetworkFetcherTests::
       << validEncodedBytes << validDimension << validTotalPixels << qint64(0);
   QTest::newRow("negative-timeout")
       << validEncodedBytes << validDimension << validTotalPixels << qint64(-1);
+  // std::chrono::milliseconds::rep is platform-defined -- typically
+  // "long" (not Qt's "long long"-based qint64) under 64-bit Linux/glibc,
+  // but "long long" under macOS/libc++. QTest::addColumn<qint64>() above
+  // requires an EXACT type match on every pushed value, so an implicit
+  // "long" here silently registers as a different, unrelated metatype
+  // and QFETCH(qint64, ...) aborts with a fatal type-mismatch assert --
+  // this only ever showed up on Linux CI, never on macOS. An explicit
+  // qint64 cast keeps the column type identical on every platform.
   QTest::newRow("timeout-above-sane-cap")
       << validEncodedBytes << validDimension << validTotalPixels
-      << (AssetNetworkFetcher::kMaxAllowedTimeout.count() + 1);
+      << static_cast<qint64>(AssetNetworkFetcher::kMaxAllowedTimeout.count() +
+                             1);
 }
 
 void AssetNetworkFetcherTests::
